+++ zone/lua_parser.cpp
@@ -1464,27 +1464,27 @@ void LuaParser::CommonOutgoingHitSuccess(Mob *self, Mob *other, DamageHitInfo &h
 	}
 }
 
-uint32 LuaParser::GetRequiredAAExperience(Client *self, bool &ignoreDefault)
+uint64 LuaParser::GetRequiredAAExperience(Client *self, bool &ignoreDefault)
 {
-	uint32 retval = 0;
+	uint64 retval = 0;
 	for (auto &mod : mods_) {
 		mod.GetRequiredAAExperience(self, retval, ignoreDefault);
 	}
 	return retval;
 }
 
-uint32 LuaParser::GetEXPForLevel(Client *self, uint16 level, bool &ignoreDefault)
+uint64 LuaParser::GetEXPForLevel(Client *self, uint16 level, bool &ignoreDefault)
 {
-	uint32 retval = 0;
+	uint64 retval = 0;
 	for (auto &mod : mods_) {
 		mod.GetEXPForLevel(self, level, retval, ignoreDefault);
 	}
 	return retval;
 }
 
-uint32 LuaParser::GetExperienceForKill(Client *self, Mob *against, bool &ignoreDefault)
+uint64 LuaParser::GetExperienceForKill(Client *self, Mob *against, bool &ignoreDefault)
 {
-	uint32 retval = 0;
+	uint64 retval = 0;
 	for (auto &mod : mods_) {
 		mod.GetExperienceForKill(self, against, retval, ignoreDefault);
 	}
